# api/openapi.yaml
openapi: 3.1.0
...
info:
  contact:
    name: Barnaby Keene
  description: >
    Storyden social API for building community driven platforms.

    The Storyden API does not adhere to semantic versioning but instead applies
    a rolling strategy with deprecations and minimal breaking changes. This has
    been done mainly for a simpler development process and it may be changed to
    a more fixed versioning strategy in the future. Ultimately, the primary way
    Storyden tracks versions is dates, there are no set release tags currently.
  title: storyden
  version: "v1.25.13-canary"
servers:
  - url: "/api"
    description: >
      The HTTP interface that this document describes is mounted on the `/api/`
      path and any requests outside of the base path will not be covered by the
      API specification and will be used for experimental features and plugins.

# All endpoints are by default accessible via a browser session cookie or an API
# key for either human accounts or bot accounts unless explicitly specified.
security:
  - browser: []
  - access_key: []
...
paths:
...
  /beacon:
    post:
      operationId: SendBeacon
      description: |
        A catch-all endpoint for tracking read states and other things that are
        not critical to the functioning of the platform. This endpoint is fire
        and forget and does not return any meaningful data. It is designed to
        be used with the `navigator.sendBeacon` API in browsers to mark things
        such as how far down a thread a member has read, or whether or not a
        Library Page has been visited recently. It may queue the work for later
        processing and is not guaranteed to be processed immediately or at all.
      security: []
      tags: [misc]
      requestBody: { $ref: "#/components/requestBodies/Beacon" }
      responses:
        default: { $ref: "#/components/responses/InternalServerError" }
        "202":
          description: Accepted

  #
  #               888               d8b
  #               888               Y8P
  #               888
  #  8888b.   .d88888 88888b.d88b.  888 88888b.
  #     "88b d88" 888 888 "888 "88b 888 888 "88b
  # .d888888 888  888 888  888  888 888 888  888
  # 888  888 Y88b 888 888  888  888 888 888  888
  # "Y888888  "Y88888 888  888  888 888 888  888
  #
...
  /accounts:
    get:
      operationId: AccountGet
      description: Get the information for the currently authenticated account.
      tags: [accounts]
      responses:
        default: { $ref: "#/components/responses/InternalServerError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "200": { $ref: "#/components/responses/AccountGetOK" }
        "304": { $ref: "#/components/responses/NotModified" }

    patch:
      operationId: AccountUpdate
      description: Update the information for the currently authenticated account.
      tags: [accounts]
      requestBody: { $ref: "#/components/requestBodies/AccountUpdate" }
      responses:
        default: { $ref: "#/components/responses/InternalServerError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "200": { $ref: "#/components/responses/AccountUpdateOK" }
...
  /assets:
    post:
      operationId: AssetUpload
      description: Upload and process a media file.
      tags: [assets]
      requestBody: { $ref: "#/components/requestBodies/AssetUpload" }
      parameters:
        - $ref: "#/components/parameters/ContentLength"
        - $ref: "#/components/parameters/AssetNameQuery"
        - $ref: "#/components/parameters/ParentAssetIDQuery"
      responses:
        default: { $ref: "#/components/responses/InternalServerError" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "200": { $ref: "#/components/responses/AssetUploadOK" }
...
components:
...
  parameters:
...
    ContentLength:
      description: Body content length in bytes.
      name: Content-Length
      in: header
      required: true
      schema:
        type: integer
        x-go-type: int64
...
    ParentAssetIDQuery:
      description: |
        For uploading new versions of an existing asset, set this parameter to
        the asset ID of the parent asset. This must be an ID and not a filename.
        This feature is used for situations where you want to replace an asset
        in its usage context, but retain the original with a way to reference it
        for features such as editable/croppable images or file version history.
      name: parent_asset_id
      in: query
      required: false
      schema:
        type: string
    AssetNameQuery:
      description: The client-provided file name for the asset.
      name: filename
      in: query
      required: false
      schema:
        type: string
...
  requestBodies:
    Beacon:
      content:
        text/plain:
          schema: { $ref: "#/components/schemas/BeaconProps" }
...
    AccountUpdate:
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AccountMutableProps" }
...
    AssetUpload:
      description: Upload a file.
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
...
  responses:
...
    NotModified:
      description: Not modified
      headers: { <<: *cache_response_headers }
...
    NotFound:
      description: Not found
    Unauthorised:
      description: Unauthorized
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
    AccountGetOK:
      description: OK
      headers: { <<: *cache_response_headers }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Account"
    AccountUpdateOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Account"
...
    AssetUploadOK:
      description: The new URL of an uploaded file.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Asset"
...
  schemas:
...
    Identifier:
      type: string
      format: xid
      # NOTE: Does not work currently with oapi-codegen because it generates a
      # new type declaration instead of a type alias so the unmarshalling fails.
      # To work around this, there are special conversion APIs in the oapi pkg.
      x-go-type: string
      # x-go-type-import:
      #   name: xid
      #   path: github.com/rs/xid
      example: "cc5lnd2s1s4652adtu50"
      description: A unique identifier for this resource.
...
    CommonProperties:
      type: object
      required:
        - id
        - createdAt
        - updatedAt
      properties:
        id: { $ref: "#/components/schemas/Identifier" }
        createdAt:
          type: string
          format: date-time
          description: The time the resource was created.
        updatedAt:
          type: string
          format: date-time
          description: The time the resource was updated.
        deletedAt:
          type: string
          format: date-time
          description: The time the resource was soft-deleted.
        misc:
          type: object
          description: Arbitrary extra data stored with the resource.
    APIError:
      type: object
      description: |
        A description of an error including a human readable message and any
        related metadata from the request and associated services.
      required: [error]
      properties:
        error:
          description: The internal error, not intended for end-user display.
          type: string
        message:
          description: A human-readable message intended for end-user display.
          type: string
        suggested:
          description: A suggested action for the user.
          type: string
        metadata:
          description: Any additional metadata related to the error.
          type: object
          additionalProperties: true
    Metadata:
      type: object
      additionalProperties: true
      description: Arbitrary metadata for the resource.
...
    AccountName:
      type: string
      description: The account owners display name.
      example: Barnaby Keene
    AccountHandle:
      type: string
      x-go-type: string
      description: The unique @ handle of an account.
      example: Southclaws
    AccountBio:
      type: string
      description: The rich-text bio for an account's public profile.
      example: <body><p>hi, my name is</p><p>southclaws</p></body>
    EmailAddress:
      description: A valid email address.
      type: string
      example: "hello@storyden.org"
    ProfileReference:
      type: object
      description: A minimal reference to an account.
      required: [id, joined, handle, name, roles]
      properties:
        id: { $ref: "#/components/schemas/Identifier" }
        joined: { $ref: "#/components/schemas/MemberJoinedDate" }
        suspended: { $ref: "#/components/schemas/MemberSuspendedDate" }
        handle: { $ref: "#/components/schemas/AccountHandle" }
        name: { $ref: "#/components/schemas/AccountName" }
...
    BeaconProps:
      x-go-type: string
      description: |
        A beacon is a lightweight reference to an object used for tracking
        purposes. It contains only the kind and ID of the object. This is mostly
        used for tracking read states of threads. But may be used for more.

        I should clarify, not the morally questionable kind of "tracking".
      type: object
      required: [k, id]
      properties:
        k: { $ref: "#/components/schemas/DatagraphItemKind" }
        id:
          type: string
          description: The identifier for the object related to tracking.

    #
    #        d8888      888               d8b
    #       d88888      888               Y8P
    #      d88P888      888
    #     d88P 888  .d88888 88888b.d88b.  888 88888b.
    #    d88P  888 d88" 888 888 "888 "88b 888 888 "88b
    #   d88P   888 888  888 888  888  888 888 888  888
    #  d8888888888 Y88b 888 888  888  888 888 888  888
    # d88P     888  "Y88888 888  888  888 888 888  888
    #
...
    Role:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonProperties"
        - $ref: "#/components/schemas/RoleProps"
...
    AccountRole:
      type: object
      allOf:
        - $ref: "#/components/schemas/Role"
        - $ref: "#/components/schemas/AccountRoleProps"
    AccountRoleList:
      type: array
      items:
        $ref: "#/components/schemas/AccountRole"
    AccountRoleProps:
      type: object
      required: [badge, default]
      properties:
        badge:
          description: |
            One role may be designated as a badge for the account. If ture, it
            should be displayed prominently on the profile or in other contexts.
          type: boolean
        default:
          description: |
            There are two built-in roles: everyone and admin, this boolean flag
            is set if this role is one of the default built-in roles.
          type: boolean
    RoleProps:
      type: object
      required: [name, colour, permissions]
      properties:
        name:
          type: string
        colour:
          type: string
        permissions:
          $ref: "#/components/schemas/PermissionList"
...
    Permission:
      type: string
      enum:
        # Posts (Threads, replies and posts)
        - "CREATE_POST"
        - "READ_PUBLISHED_THREADS"
        - "CREATE_REACTION"
        - "MANAGE_POSTS"
        - "MANAGE_CATEGORIES"
        - "CREATE_INVITATION"
        # Library (Page tree nodes)
        - "READ_PUBLISHED_LIBRARY"
        - "MANAGE_LIBRARY"
        - "SUBMIT_LIBRARY_NODE"
        # Assets
        - "UPLOAD_ASSET"
        # Events
        - "MANAGE_EVENTS"
        # Profiles (listing and viewing member profiles)
        - "LIST_PROFILES"
        - "READ_PROFILE"
        # Collections
        - "CREATE_COLLECTION"
        - "LIST_COLLECTIONS"
        - "READ_COLLECTION"
        - "MANAGE_COLLECTIONS"
        - "COLLECTION_SUBMIT"
        # Personal access keys for automation/MCP
        - "USE_PERSONAL_ACCESS_KEYS"
        # Administrative (Settings, bans, etc)
        - "MANAGE_SETTINGS"
        - "MANAGE_SUSPENSIONS"
        - "MANAGE_ROLES"
        - "MANAGE_REPORTS"
        - "VIEW_ACCOUNTS"
        # Administrator implicitly has all permissions.
        - "ADMINISTRATOR"
    PermissionList:
      type: array
      items:
        $ref: "#/components/schemas/Permission"

    #
    #        d8888          888    888
    #       d88888          888    888
    #      d88P888          888    888
    #     d88P 888 888  888 888888 88888b.
    #    d88P  888 888  888 888    888 "88b
    #   d88P   888 888  888 888    888  888
    #  d8888888888 Y88b 888 Y88b.  888  888
    # d88P     888  "Y88888  "Y888 888  888
    #
...
    Account:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonProperties"
        - $ref: "#/components/schemas/AccountCommonProps"
    AccountCommonProps:
      required:
        [
          joined,
          handle,
          name,
          roles,
          bio,
          links,
          meta,
          verified_status,
          email_addresses,
          admin,
        ]
      properties:
        joined: { $ref: "#/components/schemas/MemberJoinedDate" }
        suspended: { $ref: "#/components/schemas/MemberSuspendedDate" }
        handle:
          $ref: "#/components/schemas/AccountHandle"
        name:
          $ref: "#/components/schemas/AccountName"
        roles:
          $ref: "#/components/schemas/AccountRoleList"
        bio:
          $ref: "#/components/schemas/AccountBio"
        links:
          $ref: "#/components/schemas/ProfileExternalLinkList"
        meta:
          $ref: "#/components/schemas/Metadata"
        verified_status:
          $ref: "#/components/schemas/AccountVerifiedStatus"
        email_addresses:
          $ref: "#/components/schemas/AccountEmailAddressList"
        notifications:
          $ref: "#/components/schemas/NotificationCount"
        admin:
          type: boolean
        invited_by:
          $ref: "#/components/schemas/ProfileReference"
    AccountMutableProps:
      type: object
      properties:
        handle:
          $ref: "#/components/schemas/AccountHandle"
        name:
          $ref: "#/components/schemas/AccountName"
        bio:
          $ref: "#/components/schemas/AccountBio"
        interests:
          $ref: "#/components/schemas/TagNameList"
        links:
          $ref: "#/components/schemas/ProfileExternalLinkList"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    AccountVerifiedStatus:
      type: string
      enum: [none, verified_email]
    AccountEmailAddressList:
      description: |
        If the instance is configured to not use any email features for auth or
        transactional/content communications, this will always be empty.
      type: array
      items: { $ref: "#/components/schemas/AccountEmailAddress" }
    AccountEmailAddress:
      type: object
      required: [id, email_address, verified]
      properties:
        id: { $ref: "#/components/schemas/Identifier" }
        email_address: { $ref: "#/components/schemas/EmailAddress" }
        verified:
          description: Is the email address verified to be owned by the account?
          type: boolean
...
    NotificationCount:
      type: integer

    #
    # 8888888b.                                    888
    # 888   Y88b                                   888
    # 888    888                                   888
    # 888   d88P .d88b.  88888b.   .d88b.  888d888 888888
    # 8888888P" d8P  Y8b 888 "88b d88""88b 888P"   888
    # 888 T88b  88888888 888  888 888  888 888     888
    # 888  T88b Y8b.     888 d88P Y88..88P 888     Y88b.
    # 888   T88b "Y8888  88888P"   "Y88P"  888      "Y888
    #                    888
    #                    888
    #                    888
    #
...
    ProfileExternalLinkList:
      type: array
      items:
        $ref: "#/components/schemas/ProfileExternalLink"
    ProfileExternalLink:
      type: object
      required: [text, url]
      properties:
        text:
          type: string
        url:
          type: string
...
    MemberJoinedDate:
      type: string
      format: date-time
      description: The time the resource was created.
    MemberSuspendedDate:
      type: string
      format: date-time
      description: The time the resource was created.

    #
    #  .d8888b.           888
    #  .d8888b 888                                        888
    # d88P  Y88b888                                        888
    # 888    888888                                        888
    # 888       88888b.  8888b. 88888b. 88888b.  .d88b. 888 .d8888b
    # 888       888 "88b    "88b888 "88b888 "88bd8P  Y8b888 88K
    # 888    888888  888.d888888888  888888  88888888888888 "Y8888b.
    # Y88b  d88P888  888888  888888  888888  888Y8b.    888      X88
    #  "Y8888P" 888  888"Y888888888  888888  888 "Y8888 888  88888P'
    #
...
    TagName:
      type: string
      description: The name of a tag.
...
    TagNameList:
      type: array
      items: { $ref: "#/components/schemas/TagName" }
...
    AssetID:
      $ref: "#/components/schemas/Identifier"
...
    Asset:
      type: object
      required: [id, filename, path, mime_type, width, height]
      properties:
        id:
          $ref: "#/components/schemas/AssetID"
        filename:
          type: string
        path:
          description: |
            The API path of the asset, conforms to the schema's GET `/assets`.
          type: string
        mime_type:
          type: string
        width:
          type: number
        height:
          type: number
        # NOTE: Presence is dictated by the callee, not the API (currently.)
        parent: { $ref: "#/components/schemas/Asset" }
...
    DatagraphItemKind:
      type: string
      enum: [post, thread, reply, node, collection, profile, event]
...
  securitySchemes:
    browser:
      type: apiKey
      in: cookie
      name: storyden-session
    access_key:
      type: http
      scheme: bearer
    webauthn:
      type: apiKey
      in: cookie
      name: "storyden-webauthn-session"
#
#     ..:.                                                        .***=
#  .+#%%%%%#+   ...                                               :%%%+
#  #%%%+==+#.  =%%%.                                              :%%%+
# :%%%%:      -*%%%=-.   :==+==:    ---: :== ----     .---- .-===-:%%%+   :==+=-:   .---..-==-.
#  +%%%%%#*=. #%%%%%%-.+%%%%#%%%%+ .%%%##%%% -%%%#.  .#%%#=*%%%%%%%%%%+ =%%#+=+#%#- -%%%#%#%%%%*
#   .-+#%%%%%- +%%%:..#%%%=   +%%%*.%%%#-. .  :#%%#..#%%#:#%%%=  .+%%%+-%%%*===+%%%:-%%%+  .#%%%.
#        +%%%# =%%%. .%%%#    .%%%%.%%%+       .#%%##%%#:.%%%#    :%%%+=%%%*+++++++:-%%%:   +%%%.
# .##*===#%%%= =%%%.  +%%%#=-=#%%%-.%%%+        .#%%%%*.  +%%%#=-=#%%%+.#%%+:..=++= -%%%:   +%%%.
# -*%%%%%%%*-  =%%%.   :*%%%%%%#+: .%%%+         -%%%#.    -*%%%%%*%%%+ .+#%%%%%%+. -%%%:   +%%%.
#    .:::.                .:::.                 -%%%*.        .:.          .:::.
#                                              -%%%*
#                                             -***+
#